+++ src/commands/basicCommands/add.js
