+++ src/commands/basicCommands/cat.js
